+++ Sources/BraintreeCore/BTAPITimingDelegate.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol BTHTTPNetworkTiming: AnyObject {
+    func fetchAPITiming(path: String, startTime: Int, endTime: Int)
+}