+++ Sources/BraintreeCore/BTHTTP.swift
@@ -2,7 +2,7 @@ import Foundation
 import Security
 
 /// Performs HTTP methods on the Braintree Client API
-class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
+class BTHTTP: NSObject, NSCopying, URLSessionTaskDelegate {
 
     typealias RequestCompletion = (BTJSON?, HTTPURLResponse?, Error?) -> Void
 
@@ -14,12 +14,13 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
 
     /// An array of pinned certificates, each an NSData instance consisting of DER encoded x509 certificates
     let pinnedCertificates: [NSData] = BTAPIPinnedCertificates.trustedCertificates()
+    let baseURL: URL
 
     /// DispatchQueue on which asynchronous code will be executed. Defaults to `DispatchQueue.main`.
     var dispatchQueue: DispatchQueue = DispatchQueue.main
-    let baseURL: URL
-    let cacheDateValidator: BTCacheDateValidator = BTCacheDateValidator()
     var clientAuthorization: ClientAuthorization?
+
+    weak var networkTimingDelegate: BTHTTPNetworkTiming?
 
     /// Session exposed for testing
     lazy var session: URLSession = {
@@ -28,7 +29,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
         
         let delegateQueue: OperationQueue = OperationQueue()
         delegateQueue.name = "com.braintreepayments.BTHTTP"
-        delegateQueue.maxConcurrentOperationCount = OperationQueue.defaultMaxConcurrentOperationCount
         
         return URLSession(configuration: configuration, delegate: self, delegateQueue: delegateQueue)
     }()
@@ -99,15 +99,11 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
 
     // MARK: - HTTP Methods
 
-    func get(_ path: String, parameters: Encodable? = nil, shouldCache: Bool = false, completion: @escaping RequestCompletion) {
+    func get(_ path: String, parameters: Encodable? = nil, completion: @escaping RequestCompletion) {
         do {
             let dict = try parameters?.toDictionary()
             
-            if shouldCache {
-                httpRequestWithCaching(method: "GET", path: path, parameters: dict, completion: completion)
-            } else {
-                httpRequest(method: "GET", path: path, parameters: dict, completion: completion)
-            }
+            httpRequest(method: "GET", path: path, parameters: dict, completion: completion)
         } catch let error {
             completion(nil, nil, error)
         }
@@ -127,53 +123,7 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
         }
     }
 
-    func put(_ path: String, parameters: [String: Any]? = nil, completion: @escaping RequestCompletion) {
-        httpRequest(method: "PUT", path: path, parameters: parameters, completion: completion)
-    }
-
-    func delete(_ path: String, parameters: [String: Any]? = nil, completion: @escaping RequestCompletion) {
-        httpRequest(method: "DELETE", path: path, parameters: parameters, completion: completion)
-    }
-
     // MARK: - HTTP Method Helpers
-
-    func httpRequestWithCaching(
-        method: String,
-        path: String,
-        parameters: [String: Any]? = [:],
-        completion: RequestCompletion?
-    ) {
-        createRequest(method: method, path: path, parameters: parameters) { request, error in
-            guard let request = request else {
-                self.handleRequestCompletion(data: nil, request: nil, shouldCache: false, response: nil, error: error, completion: completion)
-                return
-            }
-
-            var cachedResponse: CachedURLResponse? = URLCache.shared.cachedResponse(for: request) ?? nil
-
-            if self.cacheDateValidator.isCacheInvalid(cachedResponse ?? nil) {
-                URLCache.shared.removeAllCachedResponses()
-                cachedResponse = nil
-            }
-
-            // The increase in speed of API calls with cached configuration caused an increase in "network connection lost" errors.
-            // Adding this delay allows us to throttle the network requests slightly to reduce load on the servers and decrease connection lost errors.
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                if let cachedResponse = cachedResponse {
-                    self.handleRequestCompletion(data: cachedResponse.data, request: nil, shouldCache: false, response: cachedResponse.response, error: nil, completion: completion)
-                } else {
-                    self.session.dataTask(with: request) { [weak self] data, response, error in
-                        guard let self else {
-                            completion?(nil, nil, BTHTTPError.deallocated("BTHTTP"))
-                            return
-                        }
-
-                        handleRequestCompletion(data: data, request: request, shouldCache: true, response: response, error: error, completion: completion)
-                    }.resume()
-                }
-            }
-        }
-    }
 
     func httpRequest(
         method: String,
@@ -183,17 +133,17 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
     ) {
         createRequest(method: method, path: path, parameters: parameters) { request, error in
             guard let request = request else {
-                self.handleRequestCompletion(data: nil, request: nil, shouldCache: false, response: nil, error: error, completion: completion)
+                self.handleRequestCompletion(data: nil, request: nil, response: nil, error: error, completion: completion)
                 return
             }
-
+            
             self.session.dataTask(with: request) { [weak self] data, response, error in
                 guard let self else {
                     completion?(nil, nil, BTHTTPError.deallocated("BTHTTP"))
                     return
                 }
 
-                handleRequestCompletion(data: data, request: request, shouldCache: false, response: response, error: error, completion: completion)
+                handleRequestCompletion(data: data, request: request, response: response, error: error, completion: completion)
             }.resume()
         }
     }
@@ -316,7 +266,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
     func handleRequestCompletion(
         data: Data?,
         request: URLRequest?,
-        shouldCache: Bool,
         response: URLResponse?,
         error: Error?,
         completion: RequestCompletion?
@@ -363,15 +312,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
                 callCompletionAsync(with: completion, body: nil, response: nil, error: error)
             }
             return
-        }
-
-        // We should only cache the response if we do not have an error and status code is 2xx
-        let successStatusCode: Bool = httpResponse.statusCode >= 200 && httpResponse.statusCode < 300
-
-        if request != nil && shouldCache && successStatusCode, let request = request {
-            let cachedURLResponse: CachedURLResponse = CachedURLResponse(response: response, data: data)
-
-            URLCache.shared.storeCachedResponse(cachedURLResponse, for: request)
         }
 
         callCompletionAsync(with: completion, body: json, response: httpResponse, error: nil)
@@ -487,7 +427,7 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
         }
     }
 
-    // MARK: - URLSessionDelegate conformance
+    // MARK: - URLSessionTaskDelegate conformance
 
     func urlSession(_ session: URLSession, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
         if challenge.protectionSpace.authenticationMethod == NSURLAuthenticationMethodServerTrust {
@@ -509,6 +449,20 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
             }
         } else {
             completionHandler(.performDefaultHandling, nil)
+        }
+    }
+
+    func urlSession(_ session: URLSession, task: URLSessionTask, didFinishCollecting metrics: URLSessionTaskMetrics) {
+        metrics.transactionMetrics.forEach { transaction in
+            if let startDate = transaction.fetchStartDate,
+               let endDate = transaction.responseEndDate,
+               let path = transaction.request.url?.path {
+                networkTimingDelegate?.fetchAPITiming(
+                    path: path,
+                    startTime: startDate.utcTimestampMilliseconds,
+                    endTime: endDate.utcTimestampMilliseconds
+                )
+            }
         }
     }
 }