+++ Sources/BraintreeCore/Date+MillisecondTimestamp.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension Date {
+    
+    var utcTimestampMilliseconds: Int {
+        Int(round(timeIntervalSince1970 * 1000))
+    }
+}